### 2025-07-28 #253 @dvklopfenstein
pmidcite
Submitting Author: DV Klopfenstein, PhD (@dvklopfenstein)
Package Name: pmidcite
One-Line Description of Package: Download "Cited by" data from the NIH for any paper with a PubMed ID from the cli
Repository Link: https://github.com/dvklopfenstein/pmidcite

---

The `pmidcite` package does the following from the cli or a script:
- **Quickly download citation data** for any paper with a PubMed ID (PMID) with minimal typing on the command line. Citation data includes citation counts (aka "CitedBy" counts) and the performance of the paper within its co-citation network.
- **Sort and sift through papers faster**
- **Make your literature search reproducible**
- **Compact information**: Download a succinct single line containing:
  - All citation and reference counts
  - Number of authors
  - First author's last name
  - Title for each researcher-given PMID by default.
  - <u>MeSH (PubMed's Medical Subject Headings) keywords</u> related to studies focusing on *human*, *animal*, and/or *molecular/cellular* biology are indicated as present if they are associated with the paper.
- **Verbose option:** Download *one information-packed line* for each:
  - The paper of interest
  - Each *citation* of the paper of interest
  - Each paper *referenced* by the paper of interest

### Published in *Research Synthesis Methods* as an invited commentary:
- **DOI:** http://dx.doi.org/10.1002/jrsm.1456
- **PubMed:** https://pubmed.ncbi.nlm.nih.gov/33031632/

We partner with communities to support peer review with an additional layer of
checks that satisfy community requirements. If your package fits into an

- [ ] Astropy: [My package adheres to Astropy community standards](https://www.pyopensci.org/software-peer-review/partners/astropy.html)

---

- Explain how and why the package falls under these categories (briefly, 1-2 sentences). For community partnerships, check also their specific guidelines as documented in the links above. Please note any areas you are unsure of:

[Conduct]: https://www.pyopensci.org/handbook/CODE_OF_CONDUCT.html
