### 2025-07-30 c5baa1a
Add pylint workflow
