+++ .github/workflows/pylint.yml
@@ -1,0 +1,43 @@
+# pylint
+# https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/controlling-permissions-for-github_token
+name: Pylint
+
+on: [push, pull_request]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    #permissions:
+      # actions: none
+      # attestations: none
+      # checks: none
+      #contents: read
+      # deployments: none
+      # id-token: write|none
+      # issues: none
+      # models: read|none
+      # discussions: none
+      #packages: read
+      # pages: none
+      # pull-requests: read
+      # security-events: none
+      # statuses: none
+    strategy:
+      matrix:
+        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v4
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v3
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install --upgrade build
+        pip install --upgrade setuptools[core]
+        pip install --upgrade pylint
+    - name: Analyzing the code with pylint
+      run: |
+        pylint $(git ls-files '*.py')
