+++ pmidcite/cli/dnldpubmed.py
@@ -7,42 +7,31 @@ import argparse
 
 from pmidcite.eutils.cmds.pubmed import PubMed
 from pmidcite.cfg import Cfg
+from pmidcite.cli.common import add_args
 from pmidcite.cli.utils import get_pmids
 
 
-class DnldPubMed:
+class BasePubMed:
     """Given PMIDs, download PubMed entries into a text file"""
+    # pylint: disable=too-few-public-methods
 
     def __init__(self):
         cfg = Cfg()
         self.cfgdct = cfg.cfgparser
         self.pubmed = PubMed(email=cfg.get_email(), apikey=cfg.get_apikey(), tool=cfg.get_tool())
 
+
+class DnldPubMed(BasePubMed):
+    """Given PMIDs, download PubMed entries into a text file"""
+
     def get_argparser(self):
         """Argument parser for Python wrapper of NIH's iCite given PubMed IDs"""
         parser = argparse.ArgumentParser(description="Run NIH's iCite given PubMed IDs")
         dir_pubmed_txt = self.cfgdct['pmidcite']['dir_pubmed_txt']
-        parser.add_argument(
-            'pmids', metavar='PMID', type=int, nargs='*',
-            help='PubMed IDs (PMIDs)')
-        parser.add_argument(
-            '-i', '--infile', nargs='*',
-            help='Read PMIDs from a file containing one PMID per line.')
-        ## parser.add_argument(
-        ##     '-o', '--outfile',
-        ##     help='Write current citation report to an ASCII text file.')
-        ## parser.add_argument(
-        ##     '-f', '--force_write', action='store_true',
-        ##     help='if an existing outfile file exists, overwrite it.')
-        parser.add_argument(
-            '-D', '--force_download', action='store_true',
-            help='Download PMID iCite information to a Python file, over-writing if necessary.')
+        add_args(parser, ['pmids', 'infile', 'force_download'])
         parser.add_argument(
             '--dir_pubmed_txt', default=dir_pubmed_txt,
-            help='Write PubMed entry into directory (default={D})'.format(D=dir_pubmed_txt))
-        ## parser.add_argument(
-        ##     '-c', '--wordcloud_filename',
-        ##     help='Output filename (i.e. pmids.png) for a word cloud plot for the given PMIDs')
+            help=f'Write PubMed entry into directory (default={dir_pubmed_txt})')
         return parser
 
     def cli(self):
@@ -51,7 +40,7 @@ class DnldPubMed:
         args = argparser.parse_args()
         # 1) Get PMIDs
         pmids = self._get_pmids(args, argparser)
-        print('{N} PMIDs'.format(N=len(pmids)))
+        print(f'{len(pmids)} PMIDs')
         # 2) Download PubMed entries.
         if pmids:
             self.pubmed.dnld_wr1_per_pmid(pmids, args.force_download, args.dir_pubmed_txt)
@@ -63,6 +52,49 @@ class DnldPubMed:
         if not pmids:
             argparser.print_help()
         return pmids
+
+
+class QueryPubMed(BasePubMed):
+    """Given PMIDs, download PubMed entries into a text file"""
+
+    def get_argparser(self):
+        """Argument parser for Python wrapper of NIH's iCite given PubMed IDs"""
+        parser = argparse.ArgumentParser(description="Run NIH's iCite given PubMed IDs")
+        dir_pubmed_txt = self.cfgdct['pmidcite']['dir_pubmed_txt']
+        parser.add_argument(
+            'query', metavar='query', type=str, nargs='*',
+            help='PubMed query')
+        parser.add_argument(
+            '-p', '--pubmed', action='store_true',
+            help='Download PubMed entry containing title, abstract, authors, journal, MeSH, etc.')
+        parser.add_argument(
[…]
