+++ pmidcite/eutils/pubmed/counts/dnld.py
@@ -2,12 +2,12 @@
 
 # To check whether a journal is indexed in MEDLINE and stored in PMC,
 # search the NLM Catalog:
-# 
+#
 #     https://www.ncbi.nlm.nih.gov/nlmcatalog
-# 
-# PMC also maintains a journals list, which is dynamic list, 
+#
+# PMC also maintains a journals list, which is dynamic list,
 # that provides journal participation level information:
-# 
+#
 #     https://www.ncbi.nlm.nih.gov/pmc/journals/
 #
 # Lidia Hutcherson
@@ -153,7 +153,7 @@ class PubMedDnld(EntrezUtilities):
     ])
 
     def __init__(self, email, apikey, tool):
-        super(PubMedDnld, self).__init__(email, apikey, tool)
+        super().__init__(email, apikey, tool)
         self.date = str(datetime.datetime.now().date()).replace('-', '_')
 
     def get_content_counts(self, file_py, force_dnld):
@@ -168,11 +168,12 @@ class PubMedDnld(EntrezUtilities):
     @staticmethod
     def load_count_data(fin_py):
         """Load NIH iCite information from Python modules"""
+        # pylint: disable=import-outside-toplevel
         import importlib
         spec = importlib.util.spec_from_file_location("module.name", fin_py)
         mod = importlib.util.module_from_spec(spec)
         spec.loader.exec_module(mod)
-        print('  READ: {PY}'.format(PY=fin_py))
+        print(f'  READ: {fin_py}')
         return mod.CNTS, mod.DATE
 
     def dnld_content_counts(self):
@@ -180,7 +181,7 @@ class PubMedDnld(EntrezUtilities):
         name2cnt = {}
         for name, query in self.name2query.items():
             cnt = self.dnld_count(query)
-            print('  Downloaded {N:12,} {A:20} {Q}'.format(N=cnt, A=name, Q=query))
+            print(f'  Downloaded {cnt:12,} {name:20} {query}')
             name2cnt[name] = cnt
         # 2019/01/04 PubMed Query returns 30,500,360 results: all [sb]
         # 2019/01/04 PubMed Query returns  3,499,063 results: medline[sb] AND pubmed pmc[sb]
@@ -211,28 +212,28 @@ class PubMedDnld(EntrezUtilities):
 
     def wrpy_count_data(self, fout_py, name2cnt):
         """Write PubMed count data into a Python module"""
-        with open(fout_py, 'w') as prt:
+        with open(fout_py, 'w', encoding='utf-8') as prt:
             prt.write('"""Downloaded PubMed count data"""\n\n')
-            prt.write("DATE = '{DATE}'\n\n".format(DATE=self.date))
+            prt.write(f"DATE = '{self.date}'\n\n")
             prt.write('# pylint: disable=line-too-long\n')
             self.prt_content_counts(name2cnt, prt)
             prt.write('\n')
             self.prt_content_cntdct(name2cnt, prt)
-            print('  WROTE: {PY}'.format(PY=fout_py))
+            print(f'  WROTE: {fout_py}')
 
     def prt_content_counts(self, name2cnt, prt=sys.stdout):
         """Print the content typename and the count of that type"""
         for name, query in self.name2query.items():
             if name in name2cnt:
                 cnt = name2cnt[name]
-                prt.write('# {N:10,} {NAME:20} {Q}\n'.format(N=cnt, NAME=name, Q=query))
+                prt.write(f'# {cnt:10,} {name:20} {query}\n')
 
[…]
