+++ pmidcite/cli/querypubmed.py
@@ -44,7 +44,7 @@ class QueryPubMed:
             help='Download PMID iCite information to a Python file, over-writing if necessary.')
         parser.add_argument(
             '--dir_pubmed_txt', default=dir_pubmed_txt,
-            help='Write PubMed entry into directory (default={D})'.format(D=dir_pubmed_txt))
+            help=f'Write PubMed entry into directory (default={dir_pubmed_txt})')
         ## parser.add_argument(
         ##     '-c', '--wordcloud_filename',
         ##     help='Output filename (i.e. pmids.png) for a word cloud plot for the given PMIDs')
