+++ pmidcite/eutils/pubmed/counts/plt.py
@@ -3,14 +3,15 @@
 __copyright__ = "Copyright (C) 2019-present, DV Klopfenstein, PhD. All rights reserved."
 __author__ = "DV Klopfenstein, PhD"
 
+# pylint: disable=wrong-import-position
 import matplotlib as mpl
 mpl.use('agg')
 import matplotlib.pyplot as plt
 
-# pylint: disable=wrong-import-position
 from pmidcite.eutils.pubmed.counts.data import DataMgr
 
 
+# pylint: disable=too-few-public-methods
 class PubMedPlot:
     """Plot the types of content and their amount in PubMed"""
 
@@ -93,50 +94,6 @@ class PubMedPlot:
         # bar text: other
         # XY-COORD [(28707594, 1819828)] (5.5, 1.8)
         axes.text(28707594+ 1819828/2.0, 5.7, 'Other', fontsize=6.3, color='black', ha='center', **dct_txt)
-
-    def chk_content_counts(self):
-        """Check the content typename and the count of that type"""
-        a2n = self.name2cnt
-        assert a2n['nihms_pub0'] + a2n['nihms_pub1'] == a2n['nihms']
-        assert a2n['pmcsd_pmc0'] + a2n['pmcsd_pmc1'] == a2n['pmcsd']
-        assert a2n['medline_pmc0'] + a2n['medline_pmc1'] == a2n['medline_all']
-        #    557,390 inprocess_A_all      inprocess[sb]
-        #    422,103 inprocess_A_pmc0     inprocess[sb] NOT pubmed pmc[sb]
-        #    135,287 inprocess_A_pmc1     inprocess[sb] AND pubmed pmc[sb]
-        #    557,390 inprocess_ml0        inprocess[sb] NOT medline[sb]
-        #          0 inprocess_ml1        inprocess[sb] AND medline[sb]
-        assert a2n['inprocess_A_all'] == a2n['inprocess_A_pmc0'] + a2n['inprocess_A_pmc1']
-        # 30,514,237 all          all [sb]
-        #  1,818,474 all_ml0_pmc0 all [sb] NOT inprocess[sb] NOT medline[sb] NOT pubmed pmc[sb]
-        # 28,695,763 ml1_pmc1     inprocess[sb] OR medline[sb] OR pubmed pmc[sb]
-        assert a2n['all'] == a2n['all_ml0_pmc0'] + a2n['ml1_pmc1']
-        #  3,116,339 pmnml_A              pubmednotmedline[sb]
-        #  3,116,339 pmnml_B              pubmednotmedline[sb] NOT medline[sb]
-        #  3,116,339 pmnml_C_ip0          pubmednotmedline[sb] NOT inprocess[sb]
-        #          0 pmnml_0_ip1          pubmednotmedline[sb] AND inprocess[sb]
-        #  1,612,274 pmnml_A_pmc1         pubmednotmedline[sb] AND pubmed pmc[sb]
-        #  1,504,065 pmnml_A_pmc0         pubmednotmedline[sb] NOT pubmed pmc[sb]
-        assert a2n['pmnml_A'] == a2n['pmnml_A_pmc1'] + a2n['pmnml_A_pmc0']
-        #  5,323,939 pmc_all              pubmed pmc[sb]
-        #    135,287 inprocess_A_pmc1     inprocess[sb] AND pubmed pmc[sb]
-        #  3,503,057 medline_pmc1         medline[sb] AND pubmed pmc[sb]
-        #  1,612,274 pmnml_A_pmc1         pubmednotmedline[sb] AND pubmed pmc[sb]
[…]
