+++ src/bin/plt_guassian_nihperc.py
@@ -9,7 +9,7 @@ import math
 import matplotlib as mpl
 mpl.use('agg')
 import matplotlib.pyplot as plt
-import scipy.stats as stats
+from scipy import stats
 from pmidcite.plot.nih_perc import PltNihVals
 from pmidcite.cfg import Cfg
 
