+++ .github/workflows/pylint.yml
@@ -40,6 +40,7 @@ jobs:
         pip install --upgrade pylint
         pip install --upgrade matplotlib
         pip install --upgrade scipy
+        pip install --upgrade requests
     - name: Analyzing the code with pylint
       run: |
         pylint $(git ls-files '*.py')

[…]
