### 2025-08-01 be676cd
make fnc private
