+++ pmidcite/cli/querypubmed.py
@@ -40,7 +40,7 @@ class QueryPubMed:
             '-p', '--pubmed', action='store_true',
             help='Download PubMed entry containing title, abstract, authors, journal, MeSH, etc.')
         parser.add_argument(
-            '-D', '--force_download', action='store_true',
+            '-f', '--force_download', action='store_true',
             help='Download PMID iCite information to a Python file, over-writing if necessary.')
         parser.add_argument(
             '--dir_pubmed_txt', default=dir_pubmed_txt,
