### 2025-08-05 #253 @eliotwrobson
Hi @dvklopfenstein thank you for your submission! pmidcite seems like a useful scientific tool, and I think that this would be a significant asset to our ecosystem. The EiC checklist is below, but outside of this, it would be great if you could fill in the blank sections on your original presubmission inquiry (the bullet points that are left blank at the bottom of your original post).

---
---

- I didn't check off the box in documentation for "short tutorials" since this is really covered by the examples in the jupyter notebooks, but these are not included directly in the documentation site. It would be best if these notebooks were integrated directly into the docs somehow.
- I didn't go through the README checks as thoroughly as in the full review, but the complete README checklist that will be used for the review is [here](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#peer-review-template) if you'd like to get started. In particular, you should make sure to include a link to the documentation.
- Could you move the contributing and code of conduct files to the base directory in the repo? It doesn't look like these are used as part of the docs site, and this made the files harder to find.
- It doesn't look like there are automated tests, linting, or typechecking set up as part of the GitHub actions workflows. It would be best if these were all set up (at minimum, you should have your tests run as part of a workflow here).
- I think there is a way to make the package usable from the command line that uses significantly less boilerplate code than is currently present, see [this article](https://til.simonwillison.net/python/uv-cli-apps) on how to do this with uv.

Aside from the above items, the core package functionality seems to be in good shape. Feel free to close this issue and open a full submission once the above items have been resolved. In the meantime, I'll get to work finding an editor for this review👍🏽
