### 2025-08-05 #253 @dvklopfenstein
Hello,

Thank you so much for the speedy and informative review.

I am currently working on adding CI in the `dev` branch, as I know this would be needed. I'll add the rest as well and open a request for a full submission ASAP.

Thank you for your time in this review.

Regards,

@dvklopfenstein
