### 2025-09-06 a402540
Handle NIH iCite response WITHOUT nih_percentile
