+++ pmidcite/icite/dnldr/pmid_dnlder.py
@@ -55,8 +55,8 @@ class NIHiCiteDownloader(NIHiCiteDownloaderBase):
             for nih_dict in nihdicts:
                 s_wrpy(pmid2foutpy[nih_dict['pmid']], nih_dict)
             s_get_group = self.nihgrouper.get_group
-            # pylint: disable=line-too-long
-            return [NIHiCiteEntry.from_jsondct(d, s_get_group(d['nih_percentile'])) for d in nihdicts]
+            s_from_jsondct = NIHiCiteEntry.from_jsondct
+            return [s_from_jsondct(d, s_get_group(d.get('nih_percentile'))) for d in nihdicts]
         return []
 
     def get_icite(self, pmid):
@@ -70,7 +70,7 @@ class NIHiCiteDownloader(NIHiCiteDownloaderBase):
                 self._wrpy(file_pmid, nih_dict)
                 return NIHiCiteEntry.from_jsondct(
                     nih_dict,
-                    self.nihgrouper.get_group(nih_dict['nih_percentile']))
+                    self.nihgrouper.get_group(nih_dict.get('nih_percentile')))
         return self.loader.load_icite(file_pmid)  # NIHiCiteEntry
 
     # -------------------------------------------------------------------------------------
