### 2025-09-06 1f190ca
`nih_percentile` now returns from NIH iCite api as a float or is not present (no longer -1)
