### 2025-09-06 fefdf01
Update to account for NIH iCite data may no longer have the following fields: cited_by_clin cited_by references provisional
