+++ pmidcite/__init__.py
@@ -2,6 +2,6 @@
 
 __copyright__ = 'Copyright (C) 2019-present, DV Klopfenstein, PhD. All rights reserved'
 __author__ = 'DV Klopfenstein, PhD'
-__version__ = '0.1.3'
+__version__ = '0.1.4'
 
 # Copyright (C) 2019-present, DV Klopfenstein, PhD. All rights reserved
