### 2025-09-06 3770738
Add upgrade option to pip install instructions
