### 2025-09-07 08325ab
Add graphic to notes
