### 2025-09-07 cdfb7bc
Simplify link to Jupyter notebooks
