### 2025-09-07 a81d735
Move "Code of conduct" and "Contributing" files to base repo dir to make them easier to find. https://github.com/pyOpenSci/software-submission/issues/253#issuecomment-3153329548
