+++ .github/ISSUE_TEMPLATE/feature_request.yaml
@@ -1,0 +1,43 @@
+name: Feature Request
+description: Suggest an idea for timetracker-csv
+title: "Feature Report"
+labels: [ ":new:", "enhancement" ]
+body:
+  - type: markdown
+    attributes:
+      value: |
+        # Feature Request
+        Hello, and thank you for reporting an issue!
+
+        Please fill out the points below, as it will make our process much easier.
+
+  - type: textarea
+    id: user-case
+    attributes:
+      label: Use Case/Motivation
+      description: What is the motivation / use case for changing the behavior?
+      placeholder: Tell us about your idea
+    validations:
+      required: true
+
+  - type: textarea
+    id: example-usage
+    attributes:
+      label: Example Usage
+      description: Please provide examples of the usage you would like to see.
+      placeholder: e.g `trk report --new-flag="fabulous new feature"`
+    validations:
+      required: true
+
+  - type: textarea
+    id: other-info
+    attributes:
+      label: Other Information
+      description: >
+        Is there anything else we should know?
+
+        (e.g. more detailed explanation, stacktraces, related
+        issues, suggestions how to fix, links for us to have context, eg.
+        stackoverflow, gitter, etc)
+    validations:
+      required: false
