### 2025-09-07 cebe6e0
Add Python version badges
