### 2025-09-09 9e5d886
Add CI testing
