### 2025-12-03 0a30f2a
Use _id rather than pmid
