### 2025-12-03 b995c54
get nih_percentile from nihdict
