+++ pmidcite/icite/dnldr/pmid_dnlder_only.py
@@ -26,7 +26,7 @@ class NIHiCiteDownloaderOnly(NIHiCiteDownloaderBase):
             ##self._prt(nihdicts)
             s_get_group = self.nihgrouper.get_group
             s_from_jsondct = NIHiCiteEntry.from_jsondct
-            return [s_from_jsondct(d, s_get_group(d['nih_percentile'])) for d in nihdicts]
+            return [s_from_jsondct(d, s_get_group(d.get('nih_percentile'))) for d in nihdicts]
         return []
 
     def get_icite(self, pmid):
@@ -36,7 +36,7 @@ class NIHiCiteDownloaderOnly(NIHiCiteDownloaderBase):
         if nih_dict:
             return NIHiCiteEntry.from_jsondct(
                 nih_dict,
-                self.nihgrouper.get_group(nih_dict['nih_percentile']))
+                self.nihgrouper.get_group(nih_dict.get('nih_percentile')))
         return None
 
     @staticmethod
