### 2025-12-03 a3976c2
Use _id rather than pmid
