### 2025-12-03 b2060af
Put NCBI Entrez command, einfo, in a try-except clause
