+++ .github/workflows/pylint.yml
@@ -33,6 +33,9 @@ jobs:
       with:
         python-version: ${{ matrix.python-version }}
     - name: Install dependencies
+    # Pkg mgrs have
+      # Give me the latest versions rather than what the developer started with
+      # Run any embedded scripts I may have
       run: |
         python -m pip install --upgrade pip
         pip install --upgrade build
