### 2025-12-03 5a562ff
Update for NIH-OCC field, author, now possibly being missing in new NIH-OCC data downloaded; Handle if author present, but no last name present
