### 2025-12-03 246f6cf
Lower default max of PMIDs from 1000 to 200 to account for NIH-OCC new limit of 200; Correct message for unavailable PMIDs; Use _id rather than pmid
