+++ pmidcite/icite/api.py
@@ -45,12 +45,14 @@ class NIHiCiteAPI:
 
     def dnld_nihdict(self, pmid):
         """Download NIH citation data for one researcher-spedified PMID. Return a corrected json"""
-        # pylint: disable=line-too-long
-        ##req_nihocc = f'{self.url_base}?{pmid}'  # v0.0.50 WAS https://icite.od.nih.gov/api/pubs?33031632
-        req_nihocc = f'{self.url_base}/{pmid}'    # v0.0.51 NOW https://icite.od.nih.gov/api/pubs/33031632
+        # The NIH-OCC changed their database format:
+        #     v0.0.50 WAS https://icite.od.nih.gov/api/pubs?33031632
+        #     v0.0.51 NOW https://icite.od.nih.gov/api/pubs/33031632
+        ##req_nihocc = f'{self.url_base}?{pmid}'
+        req_nihocc = f'{self.url_base}/{pmid}'
         rsp_json = self._send_request(req_nihocc)
-        ##print(f'AAAAAAAAAAAAAAAAAAAAAAAAAAAAAA {rsp_json}')
-        ##print(f'AAAAAAAAAAAAAAAAAAAAAAAAAAAAAA rsp_json["data"][{len(rsp_json["data"])}] = {rsp_json["data"]}')
+        ##print(f'AAAAAAAAAAAAAAAAA {rsp_json}')
+        ##print(f'AAAAAAAAAAAAAAAAA rsp_json["data"][{len(rsp_json["data"])}] = {rsp_json["data"]}')
         ##return self._adjust_json_entry(rsp_json) if rsp_json else None
         if rsp_json:
             assert 'data' in rsp_json, rsp_json
@@ -63,10 +65,14 @@ class NIHiCiteAPI:
         """Download a list of NIH citation data for given PMIDs"""
         return self._dnld_gtmax(pmids) if len(pmids) > 1000 else self._dnld_ltmax(pmids)
 
-    def _dnld_gtmax(self, pmids):
+    # NIH documentation says the max pmids is now 200:
+    #     https://support.icite.nih.gov/hc/en-us/articles/9513563045787-Bulk-Data-and-API
+    # Too many pmids shows up with this error:
+    # 413 Request Entity Too Large URL[4539]: https://icite.od.nih.gov/api/pubs?pmids=29095435, ...
+    #      def _dnld_gtmax(self, pmids, max_limit=1000):
+    def _dnld_gtmax(self, pmids, max_limit=200):
         """Run iCite on given PubMed IDs"""
         nih_dicts_all = []
-        max_limit = 1000
         pmid_list_all = pmids if isinstance(pmids, list) else list(pmids)
         num_total = len(pmids)
         # The NIH-OCC allows for a maximum of 1,000 PMIDs to be downloaded at once
@@ -74,7 +80,6 @@ class NIHiCiteAPI:
             nih_dicts_cur = self._dnld_ltmax(pmid_list_cur)
             if nih_dicts_cur:
                 nih_dicts_all.extend(nih_dicts_cur)
-            # pylint: disable=line-too-long
             print(f'NIH citation data downloaded: {len(nih_dicts_all):,} of {num_total:,}')
         return nih_dicts_all
 
@@ -82,8 +87,10 @@ class NIHiCiteAPI:
         """Download NIH citation data using a request using their API"""
         ## tic = default_timer()
         pmids_str = ','.join(str(p) for p in pmids)
-        # pylint: disable=line-too-long
-        req_nihocc = f'{self.url_base}?pmids={pmids_str}' # https://icite.od.nih.gov/api/pubs?pmids=33031632
+
+        # https://icite.od.nih.gov/api/pubs?pmids=33031632
[…]
