### 2025-12-03 1a05f89
Install local `pmidcite` package
