### 2025-12-03 fd492ab
do not run a test where a key is needed
