### 2025-12-03 ddaf5fe
Improved reporting for large downloads
