### 2025-12-03 0821ebe
Adjust print if title or year is not found; test for good runs even if NIH flds are missing
