### 2025-12-12 8053a6f
Working on JOSS paper
