+++ paper.bib
@@ -1,0 +1,126 @@
+@article{Boeker_2013,
+ author = {Boeker, Martin and Vach, Werner and Motschall, Edith},
+ title = {Google Scholar as replacement for systematic literature searches: good relative recall and precision are not enough},
+ journal = {BMC Medical Research Methodology},
+ publisher = {Springer Science and Business Media LLC},
+ year = {2013},
+ month = {October},
+ number = {1},
+ volume = {13},
+ url = {http://dx.doi.org/10.1186/1471-2288-13-131},
+ doi = {10.1186/1471-2288-13-131},
+ issn = {1471-2288}
+}
+
+@article{Gusenbauer_2025,
+ author = {Gusenbauer, Michael and Gauster, Sebastian P.},
+ title = {How to search for literature in systematic reviews and meta-analyses: A comprehensive step-by-step guide},
+ journal = {Technological Forecasting and Social Change},
+ publisher = {Elsevier BV},
+ year = {2025},
+ month = {March},
+ volume = {212},
+ pages = {123833},
+ url = {http://dx.doi.org/10.1016/j.techfore.2024.123833},
+ doi = {10.1016/j.techfore.2024.123833},
+ issn = {0040-1625}
+}
+
+@article{Hutchins_2019,
+ author = {Hutchins, B. Ian and Baker, Kirk L. and Davis, Matthew T. and Diwersy, Mario A. and Haque, Ehsanul and Harriman, Robert M. and Hoppe, Travis A. and Leicht, Stephen A. and Meyer, Payam and Santangelo, George M.},
+ title = {The NIH Open Citation Collection: A public access, broad coverage resource},
+ journal = {PLOS Biology},
+ publisher = {Public Library of Science (PLoS)},
+ year = {2019},
+ month = {October},
+ number = {10},
+ volume = {17},
+ pages = {e3000385},
+ url = {http://dx.doi.org/10.1371/journal.pbio.3000385},
+ doi = {10.1371/journal.pbio.3000385},
+ issn = {1545-7885}
+}
+
+@article{Johnson_2021,
+ author = {Johnson, Kelly and Ghezzi‐Kopel, Kate and Eldermire, Erin and Fournier, Chris and Johns, Erica and Kallaher, Amelia and Kibbee, Matthew and Miller, Christian},
+ title = {Response to Gusenbauer M, Haddaway NR “What every researcher should know about searching – Clarified concepts, search advice, and an agenda to improve finding in academia”},
+ journal = {Research Synthesis Methods},
+ publisher = {Wiley},
+ year = {2021},
+ month = {September},
+ number = {6},
+ volume = {12},
+ pages = {682--683},
+ url = {http://dx.doi.org/10.1002/jrsm.1519},
+ doi = {10.1002/jrsm.1519},
+ issn = {1759-2887}
+}
+
+@article{Klopfenstein_2020,
+ author = {Klopfenstein, D. V. and Dampier, Will},
+ title = {Commentary to Gusenbauer and Haddaway 2020: Evaluating retrieval qualities of Google Scholar and<scp>PubMed</scp>},
+ journal = {Research Synthesis Methods},
+ publisher = {Wiley},
+ year = {2020},
+ month = {October},
+ number = {2},
+ volume = {12},
+ pages = {126--135},
+ url = {http://dx.doi.org/10.1002/jrsm.1456},
+ doi = {10.1002/jrsm.1456},
+ issn = {1759-2887}
+}
+
+@article{sayers2025database,
+ author = {Sayers, Eric W and Beck, Jeffrey and Bolton, Evan E and Brister, J Rodney and Chan, Jessica and Connor, Ryan and Feldgarden, Michael and Fine, Anna M and Funk, Kathryn and Hoffman, Jinna and others, },
[…]
