+++ paper.md
@@ -1,0 +1,118 @@
+---
+title: 'pmidcite: get citations for a PubMed paper from the command line'
+tags:
+  - pubmed
+  - command-line-tool
+  - citation-downloader
+  - citation
+  - literature-review
+  - systematic-review
+  - google-scholar
+authors:
+  - name: DV Klopfenstein
+    orcid: 0000-0003-0161-7603
+    corresponding: true
+    affiliation: "1"
+affiliations:
+  - name: Department of Microbiology and Immunology, Drexel University College of Medicine, Philadelphia, PA, United States
+    index: 1
+date: 12 December 2025
+bibliography: paper.bib
+repository: https://github.com/dvklopfenstein/pmidcite
+archive_doi: 10.5281/zenodo.5172713
+---
+
+# Summary
+Viewing citation counts for a paper in the PubMed database is a common task when performing a literature search. <!--
+-->
+A scientist traces the evolution of a research question by
+performing forward and backward snowballing and 
+carrying out database searches. <!--
+-->
+Forward snowballing begins with studying the citation list of a chosen paper to find additional papers of interest.
+The researcher then selects papers from the citation list for further review by
+investigating the citation lists of all the newly selected papers,
+thus building a larger and larger snowball of information from which to synthesize knowledge.
+Backward snowballing uses lists of references rather than citations.[@Wohlin_2014] <!--
+-->
+Database searches add new papers to a literature search using free-text entered by the researcher into a search bar. <!--
+-->
+Here I present pmidcite, a CLI designed to facilitate a literature search
+using a paper or set of papers that are reposited into PubMed,
+a resource which is aimed to help scientists perform work that
+improves health for both individuals and people all over the world.
+PubMed contains millions and millions of citations
+from the biomedicine and health fields,
+and related disciplines such as
+life sciences, behavioral sciances, chemical, sciences, and bioengineering. <!--
+-->
+PubMed was developed and is maintained by
+the National Center for Biotechnology Information (NCBI), at
+the U.S. National Library of Medicine (NLM),
+located at the National Institutes of Health (NIH).[@sayers2025database] <!--
+-->
+Citation counts and other data is downloaded by pmidcite from the NIH Open Citation Collection.[@Hutchins_2019]    
+
+
+# Statement of need
+<!--
+The development of pmidcite was driven by the need to generate large-scale datasets of drug-like peptide SMILES strings for pretraining transformer-based models to predict membrane permeation from chemical structure [@feller2025peptide]. Built on the core concepts from the CycloPs [@duffy2011cyclops] method for FASTA-to-SMILES conversion, pmidcite has evolved into a stand-alone resource to support peptide-focused machine learning pipelines and peptide design workflows. While several bioinformatics toolkits exist for chemical representation and cheminformatics workflows [@ChemAxon; @o2011open; @landrum2013rdkit; @OEChem; @cock2009biopython];, many face limitations such as proprietary licensing and lack in ability to interpret or encode noncanonical amino acids (NCAAs). These constraints limit high-throughput application of sequence generation and conversion, especially for drug-like peptides containing diverse stereochemistries. In, addition, there are several python tools that focus on structure generation and cyclization [@tien2013peptidebuilder; @yang2025cyclicpeptide], however, these are not able to incorporate all necessary modifications. I used pmidcite to build a dataset of 10M peptides with NCAAs, backbone modifications, and cyclizations for pretraining a chemical language model [@feller2025peptide]. To support the community, have made pmidcite available as an open source package on PyPI, offering both command-line tools and Python functions for seamless integration into larger workflows.
+-->
+
+
+# Features
+<!--
+pmidcite offers five core command-line tools to support peptide sequence generation, conversion, modification, and analysis:
+
+- **generate-peptides:**  
[…]
