+++ .pubs/bib/Wohlin_2016.bib
@@ -1,0 +1,13 @@
+@inproceedings{Wohlin_2016,
+ author = {Wohlin, Claes},
+ title = {Second-generation systematic literature studies using snowballing},
+ publisher = {ACM},
+ year = {2016},
+ month = {June},
+ pages = {1--6},
+ url = {http://dx.doi.org/10.1145/2915970.2916006},
+ doi = {10.1145/2915970.2916006},
+ booktitle = {Proceedings of the 20th International Conference on Evaluation and Assessment in Software Engineering},
+ collection = {EASE ’16},
+ series = {EASE ’16}
+}
