+++ .pubs/bib/Johnson_2021.bib
@@ -1,0 +1,14 @@
+@article{Johnson_2021,
+ author = {Johnson, Kelly and Ghezzi‐Kopel, Kate and Eldermire, Erin and Fournier, Chris and Johns, Erica and Kallaher, Amelia and Kibbee, Matthew and Miller, Christian},
+ title = {Response to Gusenbauer M, Haddaway NR “What every researcher should know about searching – Clarified concepts, search advice, and an agenda to improve finding in academia”},
+ journal = {Research Synthesis Methods},
+ publisher = {Wiley},
+ year = {2021},
+ month = {September},
+ number = {6},
+ volume = {12},
+ pages = {682--683},
+ url = {http://dx.doi.org/10.1002/jrsm.1519},
+ doi = {10.1002/jrsm.1519},
+ issn = {1759-2887}
+}
