+++ .pubs/bib/Gusenbauer_2025.bib
@@ -1,0 +1,13 @@
+@article{Gusenbauer_2025,
+ author = {Gusenbauer, Michael and Gauster, Sebastian P.},
+ title = {How to search for literature in systematic reviews and meta-analyses: A comprehensive step-by-step guide},
+ journal = {Technological Forecasting and Social Change},
+ publisher = {Elsevier BV},
+ year = {2025},
+ month = {March},
+ volume = {212},
+ pages = {123833},
+ url = {http://dx.doi.org/10.1016/j.techfore.2024.123833},
+ doi = {10.1016/j.techfore.2024.123833},
+ issn = {0040-1625}
+}

[…]
