### 2025-12-27 3948fee
Start to rework summary to highlight CitedBy
