### 2026-02-03 #253 @dvklopfenstein
Thank you so much for getting back to me. I began work on the JOSS paper, but I would like to apply only for pyOpenSci and NOT JOSS at this time.

I have a clear idea of the content for the JOSS paper, but I want to take the time to ensure it is well-done and then get an independent editor to go over it with a fine-tooth comb, which will delay my pyOpenSci submission. So I would like to separate the two applications.

Thank you for the question and I will submit ASAP to pyOpenSci (without JOSS).
