### 2026-02-03 #253 @crhea93
Not a problem at all! In fact our work flow easily allows for that since the JOSS submission piece *only* begins after the package has been accepted by PyOpenSci.
